+++ build.config.ts
@@ -7,5 +7,5 @@ export default defineConfig({
   splitting: false,
   sourcemap: true,
   clean: true,
-  external: ["react", "@hitaqnia/core"],
+  external: ["react", "@hitaqnia/huwiya-core"],
 });